Wrote a simple argument parser.
This is perhaps the single most useless task in the world: writing
another argument parser since there is at least a dozen libraries that
do this just for Java.  However, none of them appear to provide us
with the really simple functionality we need and this parser is
perhaps as minimal as it is possible to get it.

It is also needed in case we have to deal with anticipated
legacy in which we get fed arguments from scripts that have
a lax approach to detail. Hrmph.