Rewrote option parsing.
Rewrote option parsing in preparation for merging copkg and Fire
Main classes to one class.  The reason we need to be creative
about the argument parsing is that the Fire functionality needs
to be able to get a truckload of unspecified command line
parameters.

(These can be added on using the "--" marker signifying that
the jopt-simple library should not parse anything coming after
that marker as options -- we want those to go through to our
own trivial option parser)