Added copkg shell script to run copkg.
Borrowed "mvn" script from the Maven project in order to provide a
shellscript that can be used to run copkg on the command line.  There
may be some cruft in that script for all I know, but it seems to work.
This needs testing though.

Also added tar.gz and zip binary distributions.

We still need someone with strong Windows-Fu to add a bat-file to
accomplish the same on Windows.

Since I borrowed the entire "mvn" script from Maven and rewrote it
I included the original license in that file.  I hope this is the
right thing to do.  If not, please feel free to yell at me.